Adjusted the Scheduler to run as a real MBean able to be started, managed and stopped at runtime (JMX HTML-Adaptor) etc.

+++ src/main/org/jboss/util/Schedulable.java
@@ -35,6 +35,6 @@ public interface Schedulable
     **/
    public void perform(
       Date pTimeOfCall,
-      int pRemainingRepetitions
+      long pRemainingRepetitions
    );
 }
